NJS-40-Updateing post data with user authentication

+++ rest_api_with_frontend_client/backend/src/controllers/feed.controller.ts
@@ -2,8 +2,8 @@ import { NextFunction, Request, Response } from 'express';
 import { validationResult } from 'express-validator';
 import { PostDto } from '../models/feed.model';
 import postService from '../services/post.service';
-import fileUtil from '../utils/file.util';
-import userService from '../services/user.service';
+import fileUtils from '../utils/file.utils';
+import postUtils from '../utils/post.utils';
 
 async function getPosts(req: Request, res: Response, next: NextFunction) {
   const page = Number(req.query.page || 1);
@@ -13,15 +13,7 @@ async function getPosts(req: Request, res: Response, next: NextFunction) {
     // console.log(posts);
 
     //TODO: when users added, change this lines of code
-    const postsDto = posts.posts?.map((post) => {
-      return {
-        ...post,
-        creator: {
-          _id: post.user.id,
-          name: post.user.name,
-        },
-      };
-    });
+    const postsDto = postUtils.getPostsDto(posts.posts);
 
     res.status(200).json({
       message: 'Fetched posts successfully',
@@ -30,6 +22,25 @@ async function getPosts(req: Request, res: Response, next: NextFunction) {
     });
   } catch (e) {
     next(e);
+  }
+}
+
+async function getPost(req: Request, res: Response, next: NextFunction) {
+  const postId = req.params.postId;
+
+  try {
+    const post = await postService.getPost(postId);
+
+    if (!post) {
+      throw new Error('Could not find post');
+    }
+
+    res.status(200).json({
+      message: 'Post fetched.',
+      post,
+    });
+  } catch (error) {
+    next(error);
   }
 }
 
@@ -61,46 +72,17 @@ async function createPost(req: Request, res: Response, next: NextFunction) {
       userId
     );
 
-    const user = await userService.getUserById(userId);
-
-    if (!createdPost || !user) throw new Error('Post was not created');
-
-    //TODO: when users added, change this lines of code
-    const createdPostDtm = {
-      ...createdPost,
-      creator: {
-        id: user.id,
-        name: user.name,
-      },
-    };
+    if (!createdPost || !createdPost.user) {
+      throw new Error('Post was not created');
+    }
+    const createdPostDto = postUtils.getPostDto(createdPost);
 
     res.status(201).json({
       message: 'Post created successfully',
-      post: createdPostDtm,
+      post: createdPostDto,
     });
   } catch (e) {
     next(e);
-  }
-
-  //Create post in db;
-}
-
-async function getPost(req: Request, res: Response, next: NextFunction) {
-  const postId = req.params.postId;
-
-  try {
-    const post = await postService.getPost(postId);
-
-    if (!post) {
-      throw new Error('Could not find post');
-    }
-
-    res.status(200).json({
-      message: 'Post fetched.',
-      post,
-    });
-  } catch (error) {
-    next(error);
   }
 }
 
@@ -109,6 +91,7 @@ async function updatePost(req: Request, res: Response, next: NextFunction) {
   const postId = req.params.postId;
   const { title, content } = req.body;
   let imageUrl = req.body.image;
+  const userId = req.userId!;
 
   if (!errors.isEmpty()) {
     return res.status(422).json({
@@ -130,8 +113,13 @@ async function updatePost(req: Request, res: Response, next: NextFunction) {
   try {
     const editedPost = await postService.getPost(postId);
 
+    if (editedPost?.userId !== userId) {
+      res.status(403).json({ message: 'No authorized!' });
+      return;
+    }
+
     if (editedPost && editedPost.imageUrl !== imageUrl) {
-      fileUtil.deleteFile(editedPost.imageUrl);
+      fileUtils.deleteFile(editedPost.imageUrl);
     }
 
     const updatedPost = await postService.updatePost(postId, {
@@ -140,7 +128,9 @@ async function updatePost(req: Request, res: Response, next: NextFunction) {
       imageUrl,
     });
 
-    res.status(200).json({ message: 'Post updated!', post: updatedPost });
+    const updatedPostDto = postUtils.getPostDto(updatedPost);
+
+    res.status(200).json({ message: 'Post updated!', post: updatedPostDto });
   } catch (error) {
     next(error);
   }
@@ -148,10 +138,18 @@ async function updatePost(req: Request, res: Response, next: NextFunction) {
 
 async function deletePost(req: Request, res: Response, next: NextFunction) {
   const postId = req.params.postId;
+  const userId = req.userId!;
   try {
+    const post = await postService.getPost(postId);
+
+    if (post?.userId !== userId) {
+      res.status(403).json({ message: 'No authorized!' });
+      return;
+    }
+
     const deletedPost = await postService.deletePost(postId);
 
-    if (deletedPost) fileUtil.deleteFile(deletedPost.imageUrl);
+    if (deletedPost) fileUtils.deleteFile(deletedPost.imageUrl);
 
     res.status(200).json({ message: 'Post deleted', post: deletedPost });
   } catch (error) {
